Moved helper functions to shared features Removed empty commands module package
