+++ features/embed/command.py
@@ -2,7 +2,7 @@ import logging
 import webcolors
 from .feature import is_active
 from bot import client
-from commands.helper import handle_command_exception
+from features._shared.helper import handle_command_exception
 from nextcord import Interaction, SlashOption, Permissions, Embed
 
 _logger = logging.getLogger("main")
