+++ features/hekmat/command.py
@@ -3,7 +3,7 @@ import logging
 import random
 from .feature import is_active, get_hekmat_text
 from bot import client
-from commands.helper import handle_command_exception
+from features._shared.helper import handle_command_exception
 from nextcord import Interaction, SlashOption
 
 _logger = logging.getLogger("main")
