+++ features/management/admin/settings/command.py
@@ -2,7 +2,7 @@ import data
 import features
 import logging
 from .feature import is_active
-from commands.helper import handle_command_exception
+from features._shared.helper import handle_command_exception
 from bot import client
 from nextcord import Interaction, Permissions, SlashOption
