+++ features/management/owner/subscription/command.py
@@ -1,7 +1,7 @@
 import data
 import logging
 from .feature import is_active
-from commands.helper import handle_command_exception
+from features._shared.helper import handle_command_exception
 from bot import client, ADMINS, HOME_GUILDS
 from nextcord import Interaction, Permissions, SlashOption
